Changed name of existing workflows router to old-workflows-router.js so that I can implement a new one with proper name. Added functions to node and projects model to aid helping get the correct parent node to start the workflow. Also altered the existing test workflow router to work with the new model and schema.
